Implement contextual numeric destinations

+++ tests/conformance/run/numeric-literals/lower.rs
@@ -3,7 +3,7 @@
 // Namespace: numeric-literals
 fn main() {
     let adaptive: terrane_int_support::Int = terrane_int_support::Int::from(16_i128);
-    let single: f32 = (1.5) as f32;
+    let single: f32 = 1.5_f32;
     let double: f64 = 2.25;
     let inferred: f64 = 3.5;
     let signed_value: i8 = 127;
